Unificacion estados, reestructuracion y  modificacion de UI de valoraciones
