+++ db/queries.py
@@ -117,7 +117,7 @@ def get_usuarios_by_multiple_ids(usuarios_ids):
 ## ===== Asignaturas =====
 ##========================
 
-def insert_asignatura(nombre, codigo, id_carrera, do_commit=True):
+def insert_asignatura(nombre, codigo, id_carrera = None, do_commit=True):
     cursor = get_cursor()
     try:
         cursor.execute(
